+++ matzipback/src/main/java/com/matzip/matzipback/users/command/domain/aggregate/Users.java
@@ -17,7 +17,7 @@ public class Users {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userSeq;
+    private long userSeq;
     private String userEmail;
     private String userPassword;
     private String userName;
